1b-1: the DELAY type-5 arm, and a LATENT bug it exposed in type-2/3
engine_b/eb_delay_t5.{h,c} + shim: src/master_render.c:1459-1866, the DELAY
TYPE 5 algorithm -- 408 lines, 68 coefficients, 101 state cells and FOUR ring
buffers. `null_b.py --module delay_t5` is EXACTLY 0 on all 33 scenarios. Full
both-rate runs are finishing as this is committed and are not claimed here.

IT FAILED FIRST, on exactly the four DELAY-type-5 scenarios -- the gate working
as intended. The differential showed v176 was EXACTLY 0 on every sample while
v177 tracked the port perfectly: one output's source cell was never really
written.

★ THE CAUSE IS THIS PROJECT'S OWN DOCUMENTED TRAP. IDA declares some locals
`int` when they carry float BIT PATTERNS:

    v99 = *(_DWORD *)(a1 + 10692032);      /* bit copy out of a float cell */
    *(_DWORD *)(a1 + 10692064) = v99;      /* bit copy into another        */

Transcribed literally against a float state field, BOTH assignments become
numeric CONVERSIONS and the value is destroyed twice. CLAUDE.md records this
from the original port work -- "typed locals CONVERT instead of REINTERPRET",
the 155 dual-typed cells that broke pilot 1 -- and this transcription walked
into it anyway.

★★ AND THE TYPE-2/3 ARM COMMITTED AN HOUR AGO HAS ONE TOO (v253), AND IT NULLED
EXACTLY 0 ANYWAY -- purely because the coefficient it carries holds a value the
double conversion happens to round-trip. A green gate hiding a real defect
behind the data. It was only found because type-5 has a carrier whose value
does NOT round-trip. Both modules are regenerated and both now null EXACTLY 0.

THE GUARD: arm_xform.carriers() identifies them mechanically. A local is a bit
carrier when EVERY use of it is a bare cell load or a bare cell store, with no
arithmetic anywhere -- a test that correctly excludes the ring indices, since
those ARE used in arithmetic. Carriers are emitted as `float`, so both copies
stay bit-exact. `arm_xform.py <lo> <hi>` now prints the carriers for a range
before anyone transcribes it.

Teeth anchors and brackets added for both arms.

+++ engine_b/eb_delay_t23.c
@@ -19,7 +19,7 @@ void eb_dly23_tick(eb_dly23_state *s, const eb_dly23_coef *c,
     float v58;
     float v176;
     float v177;
-    int v253;
+    float v253;
     float v254;
     float v255;
     float v256;
